% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter_plot}
\usage{
  scatter_plot(data, x, y)
}
\arguments{
  \item{data}{The input dataset}
  
  \item{x}{A continuous numerical variable in the dataset, it is one of the variable
    being investigated}
  
  \item{x_axis_name}{This is the name of the x-axis given by the user who uses this function}
}
\value{
  A scatter plot with value of x in x-axis y in y-axis
}
\description{
  The AssignmentB2AS package generates a scatter plot with ggplot2. Using selected data sets, this package offers a quick way to assess if there is a relationship between independent (x) and dependent (y) variables with the visualization options of bullet size 1 and purple color.
  
}

\examples{
  scatter_plot(x = datateachr::cancer_sample$perimeter_mean, y = datateachr::cancer_sample$texture_mean)
}