% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter_plot}
\usage{
  scatter_plot(data, x, fill, x_axis_name)
}
\arguments{
  \item{data}{The input dataset}
  
  \item{x}{A continuous numerical variable in the dataset, it is one of the variable
    being investigated}
  
  \item{fill}{A variable in the dataset, it is one of the variable being investigated}
  
  \item{x_axis_name}{This is the name of the x-axis given by the user who uses this function}
}
\value{
  A scatter plot with value of x in x-axis across different fills and scatter as the y-axis
}
\description{
  The scatter_plot function creates a scatter plot wiht alpha=0.5
  based on the given inputs, dataset, x, and fill values, using ggplot2.
  This function can be used as a visualization tool to represent the
  distribution of a continuous numeric variable (x) across multiple fills(categorical values).
}
\examples{
  scatter_plot(x = datateachr::cancer_sample$perimeter_mean, y = cancer_sample$texture_mean)
}