% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R

\name{scatter_plot}
\alias{scatter_plot}
\title{Generate a scatter plot.}
\usage{
  scatter_plot(data, x, y, x_axis_name = "X-axis", y_axis_name = "Y-axis", title = "Scatter Plot")
}
\arguments{
  \item{data}{
    The input dataset.
  }
  
  \item{x}{
    A continuous numerical variable in the dataset, representing one of
    the variables being investigated.
  }
  
  \item{y}{
    Another continuous numerical variable in the dataset, representing the
    variable being assessed for dependence on 'x'.
  }
  
  \item{x_axis_name}{
    The name for the x-axis, provided by the user calling this function.
  }
  
  \item{y_axis_name}{
    The name for the y-axis, provided by the user calling this function.
  }
  
  \item{title}{
    The title for the scatter plot.
  }
}
\value{
  A scatter plot with 'x' on the x-axis and 'y' on the y-axis.
}
\description{
  This function creates a scatter plot using ggplot2, allowing users to assess
  relationships between independent (x) and dependent (y) variables. The plot
  uses a bullet size of 1 and is colored purple.
}
\examples{
  scatter_plot(data = datateachr::cancer_sample,
               x = "perimeter_mean",
               y = "texture_mean",
               x_axis_name = "Perimeter",
               y_axis_name = "Texture",
               title = "Scatter Plot Example")
}
